requirements.txt
aiogram==3.4.1
requests
python-dotenv
google-genai